fastapi/src/requirements.txt
fastapi==0.110.0
python-dotenv==1.0.1
uvicorn==0.28.0
gunicorn==21.2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
pillow==10.2.0
pydantic==2.6.4
pydantic_core==2.16.3
requests==2.31.0
python-multipart==0.0.9
pymongo==4.6.2
motor==3.3.2